Refactor: Generic ElementPosition util + fold evaluateWhen into utils.ts [ED-24815]
- Moves ChildPosition/ChildPositionKind + resolveInsertIndex out of the
  children-dependencies module into sync/ as ElementPosition/ElementPositionKind,
  mirroring the PHP-side Element_Position move (addresses review feedback: element
  insertion positioning isn't dependency-specific).
- Folds the one-line evaluateWhen wrapper from evaluate-when.ts into utils.ts,
  removing the standalone file per review feedback.
