+++ packages/packages/libs/editor-elements/src/children-dependencies/index.ts
@@ -1,5 +1,4 @@
 export { bindSettingsReconcile } from './bind-settings-reconcile';
-export { evaluateWhen } from './evaluate-when';
+export { evaluateWhen } from './utils';
 export { reconcileInitialChildren } from './reconcile-initial-children';
-export { resolveInsertIndex } from './resolve-insert-index';
-export type { ChildDependenciesConfig, ChildDependencyRule, ChildPosition, ChildPositionKind } from './types';
+export type { ChildDependenciesConfig, ChildDependencyRule } from './types';
