+++ packages/packages/libs/editor-elements/src/children-dependencies/reconcile-initial-children.ts
@@ -1,8 +1,8 @@
+import { resolveInsertIndex } from '../sync/resolve-insert-index';
 import { type V1ElementConfig, type V1ElementData, type V1ElementSettingsProps } from '../sync/types';
-import { evaluateWhen } from './evaluate-when';
-import { resolveInsertIndex } from './resolve-insert-index';
 import { createChildrenStash } from './stash';
 import { type ChildDependencyRule } from './types';
+import { evaluateWhen } from './utils';
 
 type ReconcileInitialChildrenArgs = {
 	elementId: string;
