+++ packages/packages/libs/editor-elements/src/children-dependencies/utils.ts
